gae/standard_requirements.txt
flask
six==1.10.0
pyasn1
pyasn1_modules
google-auth
google-api-python-client
google-auth-httplib2
pyyaml
WebTest
